# api_backend/openapi.yaml
openapi: 3.0.0
info:
  title: Music App API
  version: "1.0.0"
  description: API cho ứng dụng nghe nhạc — Auth, Songs, Artists, Genres, Playlists, Favorites, Histories, Admin
  contact:
    name: API Support
    email: support@musicapp.com
servers:
  - url: http://localhost:5000
    description: Development server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
          format: email
        full_name:
          type: string
          nullable: true
        avatar_url:
          type: string
          nullable: true
        role:
          type: string
          enum: [user, artist, admin]
        status:
          type: integer
          description: 1 active, 0 locked
        created_at:
          type: string
          format: date-time
      required: [id, username, email, role]
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        errors:
          type: array
          items:
            type: object
...
    GenreDetailResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: "#/components/schemas/Genre"
...
    AvatarUploadResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Avatar updated
        avatar_url:
          type: string
          example: http://localhost:5000/uploads/avatars/avatar_default.jpg
paths:
...
  /api/register:
    post:
      tags: [Auth]
      summary: Đăng ký tài khoản mới
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, email, password, full_name]
              properties:
                username:
                  type: string
                  minLength: 3
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 6
                full_name:
                  type: string
      responses:
        "201":
          description: Đăng ký thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  user:
                    $ref: "#/components/schemas/User"
        "400":
          description: Dữ liệu không hợp lệ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Username hoặc email đã tồn tại
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/login:
    post:
      tags: [Auth]
      summary: Đăng nhập người dùng
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Đăng nhập thành công
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Thiếu thông tin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Sai username hoặc password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/me:
    get:
      tags: [Auth]
      summary: Lấy thông tin user hiện tại
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Thông tin user
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  user:
                    $ref: "#/components/schemas/User"
        "401":
          description: Không có token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/profile:
    get:
      tags: [Profile]
      summary: Lấy thông tin hồ sơ cá nhân
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Thông tin profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Profile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    put:
      tags: [Profile]
      summary: Cập nhật hồ sơ cá nhân
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                full_name:
                  type: string
                date_of_birth:
                  type: string
                  format: date
                gender:
                  type: string
                phone:
                  type: string
                address:
                  type: string
                bio:
                  type: string
      responses:
        "200":
          description: Cập nhật thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Profile"
        "400":
          description: Dữ liệu không hợp lệ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/profile/avatar:
    put:
      tags: [Profile]
      summary: Upload avatar cho profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [avatar]
              properties:
                avatar:
                  type: string
                  format: binary
      responses:
        "200":
          description: Upload avatar thành công
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvatarUploadResponse"
        "400":
          description: Không có file hoặc file không hợp lệ
        "401":
          description: Unauthorized

    delete:
      tags: [Profile]
      summary: Reset avatar về mặc định
      description: Xóa avatar hiện tại và quay về avatar mặc định
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Reset avatar thành công
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvatarUploadResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/songs:
    get:
      tags: [Songs]
      summary: Lấy danh sách bài hát với phân trang và filter
      parameters:
        - in: query
          name: keyword
          schema:
            type: string
        - in: query
          name: genreId
          schema:
            type: integer
        - in: query
          name: artistId
          schema:
            type: integer
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Danh sách bài hát
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SongsListResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/songs/{id}:
    get:
      tags: [Songs]
      summary: Lấy chi tiết bài hát và tăng view count
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chi tiết bài hát
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Song"
        "404":
          description: Không tìm thấy bài hát
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/albums:
    get:
      tags: [Albums]
      summary: Lấy danh sách album
      parameters:
        - in: query
          name: keyword
          schema:
            type: string
        - in: query
          name: artistId
          schema:
            type: integer
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Danh sách album
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlbumsListResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/albums/{id}:
    get:
      tags: [Albums]
      summary: Lấy chi tiết album
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chi tiết album
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Album"
        "404":
          description: Không tìm thấy album
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/search:
    get:
      tags: [Search]
      summary: Tìm kiếm tổng hợp (nghệ sĩ, album, bài hát)
      parameters:
        - in: query
          name: keyword
          required: true
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Kết quả tìm kiếm
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/artists:
    get:
      tags: [Artists]
      summary: Lấy danh sách nghệ sĩ
      parameters:
        - in: query
          name: keyword
          schema:
            type: string
      responses:
        "200":
          description: Danh sách nghệ sĩ
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Artist"
  /api/artists/{id}:
    get:
      tags: [Artists]
      summary: Lấy chi tiết nghệ sĩ
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chi tiết nghệ sĩ
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Artist"
        "404":
          description: Không tìm thấy nghệ sĩ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/genres/{id}:
    get:
      tags:
        - Genres
      summary: Lấy chi tiết thể loại
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        "200":
          description: Thành công
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenreDetailResponse"
        "404":
          description: Không tìm thấy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/playlists:
    get:
      tags: [Playlists]
      summary: Lấy playlists của user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Danh sách playlists
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Playlist"
    post:
      tags: [Playlists]
      summary: Tạo playlist mới
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
                isPublic:
                  type: boolean
      responses:
        "201":
          description: Playlist được tạo
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Playlist"
        "400":
          description: Dữ liệu không hợp lệ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/playlists/{id}:
    get:
      tags: [Playlists]
      summary: Lấy chi tiết playlist và danh sách bài hát
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chi tiết playlist
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/PlaylistDetail"
        "404":
          description: Không tìm thấy playlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [Playlists]
      summary: Xóa playlist (soft delete)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Kết quả xóa
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
  /api/playlists/{id}/songs:
    post:
      tags: [Playlists]
      summary: Thêm bài hát vào playlist
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [songId]
              properties:
                songId:
                  type: integer
      responses:
        "200":
          description: Kết quả thêm bài hát
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
...